[fix] post 자료형 일부 변경

+++ Mallery4/app/src/main/java/com/example/mallery4/datamodel/PostWriteResponse.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class PostWriteResponse (
     @SerializedName("state") var state:Int,
     @SerializedName("result") var result:String,
-    @SerializedName("postId") var postId:Int,
+    @SerializedName("postId") var postId:Long,
     @SerializedName("postDate") var postDate:String,
     @SerializedName("postLocation") var postLocation:String,
     @SerializedName("userId") var userId:String,
